EUI-2072-call-real-apis
