+++ api/caseshare/routes.ts
@@ -10,7 +10,6 @@ router.get('/db', restAPI.getDB)
 router.get('/orgs', restAPI.getOrgs)
 router.get('/orgs/:orgId/users', restAPI.getUsersByOrgId)
 router.get('/orgs/:orgId/users/:uid', restAPI.getUserByOrgAndUserId)
-// router.get('/orgs/:orgId', restAPI.searchUsers)
-router.get('/refdata/organisations/users', restAPI.getUsers)
+router.get('/users', restAPI.getUsers)
 router.get('/cases', restAPI.getCases)
 router.get('/cases/:caseId', restAPI.getCaseById)
