--- src/app/components/exui-mian-wrapper/exui-page-wrapper.component.html
@@ -1,14 +1,0 @@
-<div class="page">
-  <a *ngIf="backLink" [routerLink]="backLink" class="govuk-back-link">Back</a>
-  <div class="govuk-width-container">
-    <main id="content" role="main" class="govuk-main-wrapper">
-      <div class="govuk-grid-row">
-        <h1 *ngIf="title" class="govuk-heading-xl">
-          <span *ngIf="fnTitle" class="govuk-caption-xl">{{fnTitle}}</span>{{title}}</h1>
-        <ng-content>
-
-        </ng-content>
-      </div>
-    </main>
-  </div>
-</div>
