--- src/app/components/exui-mian-wrapper/exui-page-wrapper.component.scss
@@ -1,3 +1,0 @@
-.page{
-  width:100%;
-}
