+++ src/app/components/index.ts
@@ -1,6 +1,5 @@
 import { AccessibilityComponent } from './accessibility/accessibility.component';
 import { CookiePolicyComponent } from './cookie-policy/cookie-policy.component';
-import {ExuiPageWrapperComponent} from './exui-mian-wrapper/exui-page-wrapper.component';
 import { GetHelpComponent } from './get-help/get-help.component';
 import {HeaderComponent} from './header/header.component';
 import {HmctsGlobalFooterComponent} from './hmcts-global-footer/hmcts-global-footer.component';
@@ -13,7 +12,6 @@ import {ServiceDownComponent} from './service-down/service-down.component';
 import { SignedOutComponent } from './signed-out/signed-out.component';
 
 export const components: any[] = [
-  ExuiPageWrapperComponent,
   HmctsGlobalFooterComponent,
   HeaderComponent,
   HmctsGlobalHeaderComponent,
@@ -28,7 +26,6 @@ export const components: any[] = [
   SignedOutComponent
 ];
 
-export * from './exui-mian-wrapper/exui-page-wrapper.component';
 export * from './hmcts-global-footer/hmcts-global-footer.component';
 export * from './header/header.component';
 export * from './hmcts-global-header/hmcts-global-header.component';
